+++ game/modsupport.py
@@ -2,8 +2,9 @@ from typing import Type
 
 from dcs.helicopters import HelicopterType, helicopter_map
 from dcs.planes import PlaneType, plane_map
-from dcs.unittype import VehicleType
+from dcs.unittype import VehicleType, ShipType
 from dcs.vehicles import vehicle_map
+from dcs.ships import ship_map
 
 
 def helicoptermod(helicopter: Type[HelicopterType]) -> Type[HelicopterType]:
@@ -19,3 +20,8 @@ def planemod(plane: Type[PlaneType]) -> Type[PlaneType]:
 def vehiclemod(vehicle: Type[VehicleType]) -> Type[VehicleType]:
     vehicle_map[vehicle.id] = vehicle
     return vehicle
+
+
+def shipmod(ship: Type[ShipType]) -> Type[ShipType]:
+    ship_map[ship.id] = ship
+    return ship