+++ game/theater/start_generator.py
@@ -71,6 +71,7 @@ class ModSettings:
     frenchpack: bool = False
     high_digit_sams: bool = False
     ov10a_bronco: bool = False
+    swedishmilitaryassetspack: bool = False
 
 
 class GameGenerator: